Use a global setup to create the database
